+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 #### Core
 
 - **Breaking Change**: in interactive mode or ``-c`` mode without shell-name, toplevel ``$0`` indicates ``argv[0]``
+- **Breaking Change**: unbalanced brace expansions are semantic error
 - show stack trace of ignored exceptions within finally/defer block
 
 #### Builtin